+++ src/App/backend/test/Altinn.App.Integration.Tests/_testapps/basic/_scenarios/workflow-engine-success/services/SuccessfulServiceTask.cs
@@ -3,16 +3,16 @@ using Altinn.App.Core.Features.Process;
 using Microsoft.Extensions.DependencyInjection;
 using TestApp.Shared;
 
-namespace Altinn.App.Integration.Tests.Scenarios.WorkflowEngineSuccessWithoutAutoAdvance;
+namespace Altinn.App.Integration.Tests.Scenarios.WorkflowEngineSuccess;
 
-public sealed class ManualServiceTask : IServiceTask
+public sealed class SuccessfulServiceTask : IServiceTask
 {
     public string Type => "write";
 
     public Task<ServiceTaskResult> Execute(ServiceTaskContext context)
     {
-        SnapshotLogger.LogInfo("IServiceTask.Execute.SuccessWithoutAutoAdvance");
-        ServiceTaskResult result = ServiceTaskResult.SuccessWithoutAutoAdvance();
+        SnapshotLogger.LogInfo("IServiceTask.Execute.Success");
+        ServiceTaskResult result = ServiceTaskResult.Success();
         return Task.FromResult(result);
     }
 }
@@ -21,6 +21,6 @@ public static class ServiceRegistration
 {
     public static void RegisterServices(IServiceCollection services)
     {
-        services.AddTransient<IServiceTask, ManualServiceTask>();
+        services.AddTransient<IServiceTask, SuccessfulServiceTask>();
     }
 }